+++ internal/runners/secrets/secrets_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/environment"
 	"github.com/ActiveState/cli/internal/failures"
+	"github.com/ActiveState/cli/internal/runners/secrets"
 	"github.com/ActiveState/cli/internal/testhelpers/httpmock"
 	"github.com/ActiveState/cli/internal/testhelpers/osutil"
 	"github.com/ActiveState/cli/internal/testhelpers/secretsapi_test"
@@ -21,7 +22,6 @@ import (
 	secretsapi "github.com/ActiveState/cli/pkg/platform/api/secrets"
 	authMock "github.com/ActiveState/cli/pkg/platform/authentication/mock"
 	"github.com/ActiveState/cli/pkg/projectfile"
-	"github.com/ActiveState/cli/state/secrets"
 )
 
 type VariablesCommandTestSuite struct {
@@ -44,7 +44,7 @@ func (suite *VariablesCommandTestSuite) BeforeTest(suiteName, testName string) {
 	// support test projectfile access
 	root, err := environment.GetRootPath()
 	suite.Require().NoError(err, "Should detect root path")
-	err = os.Chdir(filepath.Join(root, "state", "secrets", "testdata"))
+	err = os.Chdir(filepath.Join(root, "internal", "runners", "secrets", "testdata"))
 	suite.Require().NoError(err, "Should chdir")
 
 	secretsClient := secretsapi_test.NewDefaultTestClient("bearing123")
