--- state/testdata/activestate.yaml
@@ -1,2 +1,0 @@
-name: project
-owner: someone
