--- state/testdata/state.bat
@@ -1,2 +1,0 @@
-@echo off
-echo "OUTPUT--%*--OUTPUT"
