--- state/testdata/state.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-echo "OUTPUT--$@--OUTPUT"
