--- state/testdata/withversion/activestate.yaml
@@ -1,4 +1,0 @@
-name: project
-owner: someone
-branch: master
-version: 1.2.3-123
